Move tests into test folder
